requirements.txt
flask==2.3.3
werkzeug==2.3.7
langchain==0.0.267
python-dotenv==0.19.0
boto3==1.28.0
chromadb==0.4.0
pypdf==3.9.0
tiktoken==0.5.1
unstructured==0.10.30
Flask==2.3.3
langchain==0.0.335
ollama==0.1.7
chromadb==0.4.18
sentence-transformers==2.2.2
PyPDF2==3.0.1
python-dotenv==1.0.0
boto3==1.28.0
botocore==1.31.85
s3transfer==0.6.2

# Enhanced dependencies for optimization
redis==4.5.4
flask-limiter==2.9.2
flask-caching==2.0.2
prometheus-client==0.17.1
structlog==23.1.0
python-multipart==0.0.6

# Development and monitoring
sentry-sdk[flask]==1.32.0
watchdog==3.0.0

# Additional document support
python-docx==0.8.11
python-magic==0.4.27

# Performance monitoring
psutil==5.9.5
torch==2.0.1
transformers==4.32.1